Fix dependency cycle
